Remove multiple Sensor base classes and leave on the Sensor class, as all Sensors now require platforms.
Passive test failing, but will be addressed in the next commit.
